+++ requirements.txt
@@ -1,8 +1,6 @@
-polars==0.20.28
-tqdm==4.64.1
-polars==0.20.27
-python-dateutil==2.8.2
-tqdm==4.64.1
-polars==0.20.27
-python-dateutil==2.8.2
-pytz version==2022.7.1
+pandas==2.2.3
+polars==1.14.0
+python-dateutil==2.9.0.post0
+pytz==2024.2
+rdflib==7.1.1
+tqdm==4.66.5
